// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const CategoriaDocente = {
  PRINCIPAL: 'PRINCIPAL',
  ASOCIADO: 'ASOCIADO',
  AUXILIAR: 'AUXILIAR',
  JEFE_PRACTICA: 'JEFE_PRACTICA'
} as const

export type CategoriaDocente = (typeof CategoriaDocente)[keyof typeof CategoriaDocente]


export const TipoDocente = {
  NOMBRADO: 'NOMBRADO',
  CONTRATADO: 'CONTRATADO'
} as const

export type TipoDocente = (typeof TipoDocente)[keyof typeof TipoDocente]


export const TipoAula = {
  TEORIA: 'TEORIA',
  LABORATORIO: 'LABORATORIO'
} as const

export type TipoAula = (typeof TipoAula)[keyof typeof TipoAula]


export const DiaSemana = {
  LUNES: 'LUNES',
  MARTES: 'MARTES',
  MIERCOLES: 'MIERCOLES',
  JUEVES: 'JUEVES',
  VIERNES: 'VIERNES'
} as const

export type DiaSemana = (typeof DiaSemana)[keyof typeof DiaSemana]


export const TipoAsignacion = {
  TEORIA: 'TEORIA',
  LABORATORIO: 'LABORATORIO'
} as const

export type TipoAsignacion = (typeof TipoAsignacion)[keyof typeof TipoAsignacion]


export const TipoRestriccion = {
  NO_DISPONIBLE: 'NO_DISPONIBLE',
  PREFERENCIA: 'PREFERENCIA'
} as const

export type TipoRestriccion = (typeof TipoRestriccion)[keyof typeof TipoRestriccion]


export const EstadoSesion = {
  PROGRAMADA: 'PROGRAMADA',
  EN_CURSO: 'EN_CURSO',
  FINALIZADA: 'FINALIZADA'
} as const

export type EstadoSesion = (typeof EstadoSesion)[keyof typeof EstadoSesion]


export const EstadoTurno = {
  PENDIENTE: 'PENDIENTE',
  EN_TURNO: 'EN_TURNO',
  COMPLETADO: 'COMPLETADO',
  AUSENTE: 'AUSENTE'
} as const

export type EstadoTurno = (typeof EstadoTurno)[keyof typeof EstadoTurno]